Implemented feature to open and modify existing notes
Changes:
Added click listeners on the recyclerview items which opens up the note fragment along with the content of the clicked item.

Log:
I literally spent hours trying to figure out how to set up the fragment so that the edit views will contain the content of the opened note. Then I figure that I forgot to set the binding of the viewmodel as such, this simple line of code ``` binding.noteViewModel = viewModel ```. I seriously need a mentor to help me with such problems hahaha.